+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added a `v3::StatMsg` record with an expanded 64-bit `quantity` field
 - Added `kDbnVersion` constants to each version namespace: `v1`, `v2`, and `v3`
 - Added `kUndefStatQuantity` constants to each version namespace
+- Added new off-market publishers for Eurex, and European Energy Exchange (EEX)
 
 ## 0.34.2 - 2025-05-06